scaffold client and invoice

+++ db/migrate/20171113232508_create_clients.rb
@@ -2,8 +2,8 @@ class CreateClients < ActiveRecord::Migration[5.1]
   def change
     create_table :clients do |t|
       t.string :name
-      t.decimal :age
-      t.decimal :hours
+      t.integer :age
+      t.integer :hours
       t.string :email
 
       t.timestamps
